ai_module/requirements.txt
# RAG系统依赖
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0

# 文档处理（可选）
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
